backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
requests==2.31.0
redis==5.0.1
slowapi==0.1.9
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
email-validator==2.1.0 